Fixed npe when resetting layout for task group

+++ KalypsoAFGUIFramework/src/org/kalypso/afgui/internal/TaskPerspectiveStore.java
@@ -432,8 +432,11 @@ class TaskPerspectiveStore
 
     /* Delete user definition for the current task */
     final ITask activeTask = m_executor.getActiveTask();
-    final File userFile = getUserFile( activeTask );
-    FileUtils.deleteQuietly( userFile );
+    if( activeTask != null )
+    {
+      final File userFile = getUserFile( activeTask );
+      FileUtils.deleteQuietly( userFile );
+    }
 
     /* Reset definition for the active task */
     resetSystemSettings( (PerspectiveDescriptor) perspective );
